refactor(models): promote models/asset.py to a models/asset/ package (#161)
models/asset.py had grown to ~500 lines spanning every OAM asset shape.
Split it into one module per cohesive OAM concern, with the package __init__
as the public re-export surface so 'from models.asset import X' and the
'from models import X' re-export keep working across every consumer
unchanged - no import sites touched outside the package.

- vuln.py (VulnProperty), certificate.py (TLSCertificate), endpoint.py
  (Endpoint / EndpointPage / LlmEndpoint), host.py (HostRole / HostPriority /
  HostInsight / HostScore / OpenPortsMap), service.py (Service / Product /
  ProductRelease), ip.py (IpAsset), graph.py (AttackGraph).
- Intra-package imports form a DAG (vuln / certificate leaves -> endpoint /
  host / service / ip -> graph), so no model_rebuild is needed.
- tests/models/test_asset.py mirrored into tests/models/asset/test_*.py, one
  per submodule, per the repo's test-mirror convention (importlib import-mode
  needs no __init__.py and tolerates the duplicate basenames).

@cyber_tool args_schema types (VulnProperty / AttackTree / etc.) resolve via
the same re-export, so every tool interface stays Pydantic-validated through
its args_schema - confirmed by the closed-world + args-schema contract tests
and an import smoke. The cybersquad-models skill glob (*/models/*.py, * crosses
/ in bash case) already matches the nested modules, so the hook still fires.

Pure relocation - no model shape or behaviour change.
